requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
datasets>=2.14.0
Pillow>=9.0.0
tqdm>=4.65.0
numpy>=1.24.0
ultralytics>=8.0.0
gradio>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tabulate>=0.9.0
